Revert the CCDF and Medicaid moves to redo them with per-program directories
