+++ programs/framework/tests/test_registry.py
@@ -30,7 +30,6 @@ from programs.framework.registry import (
     is_abstract,
     register,
 )
-from programs.programs.cross_white_label.medicaid.base import Medicaid
 
 
 class WalkClassesTests(SimpleTestCase):
@@ -182,8 +181,7 @@ class RegistryCoversEveryCalculatorTests(SimpleTestCase):
         harmless while one state uses it and breaks when a second arrives, because
         the base and the new state's subclass then claim the same code.
         """
-        from programs.programs.cross_white_label.ccdf.base import Ccdf
-        from programs.programs.federal.pe.member import EarlyHeadStart, HeadStart, Medicaid
+        from programs.programs.federal.pe.member import Ccdf, EarlyHeadStart, HeadStart, Medicaid
         from programs.programs.federal.pe.tax import Cdcc
 
         for base in (Cdcc, Medicaid, HeadStart, EarlyHeadStart, Ccdf):
@@ -259,8 +257,7 @@ class AbstractDeclarationTests(SimpleTestCase):
         )
 
     def test_the_family_bases_declare_themselves_abstract(self):
-        from programs.programs.cross_white_label.ccdf.base import Ccdf
-        from programs.programs.federal.pe.member import Chip, EarlyHeadStart, HeadStart, Medicaid, Msp
+        from programs.programs.federal.pe.member import Ccdf, Chip, EarlyHeadStart, HeadStart, Medicaid, Msp
         from programs.programs.federal.pe.tax import Aca, Cdcc
 
         for cls in (Medicaid, Chip, HeadStart, EarlyHeadStart, Ccdf, Msp, Aca, Cdcc):
