+++ programs/programs/federal/pe/tax.py
@@ -1,6 +1,6 @@
+from programs.programs.federal.pe.member import Medicaid
 from programs.framework.pe_base import PolicyEngineTaxUnitCalulator
 import programs.framework.pe_dependencies as dependency
-from programs.programs.cross_white_label.medicaid.base import Medicaid
 
 
 class Eitc(PolicyEngineTaxUnitCalulator):
